Implement changes to dinosaur mechanics
Stegosaurs and allosaurs are now depicted by 's' and 'a' respectively. This allows the player to distinguish between the different types of dinosaurs (previously, all dinosaurs were depicted by 'd'). Threshold for hungry dinosaurs was dropped from 50 to 30 (so dinosaurs aren't hungry as soon as the game starts). Fixed a bug where dinosaurs stopped moving as soon as they were hungry - now they stop moving if their hunger level is 0.