Implement FollowBehaviour between dinosaurs
This was achieved via the getMate method, which uses a similiar process as the querySurroundings method in the ScanSurrounds class. However, it has to search all surrounding tiles up to 3 tiles from a given location (making it computationally expensive). At the moment, the sex and type of dinosaurs is not accounted for. Also changed hunger threshold back to 50 (previously misread specification).