Implement ability for player to try and pick a fruit from a tree
Used HarvestAction (as picking fruit is a form of harvesting). Overrode the allowableActions method from the Ground class, and reimplemented it in Tree (which instantiates HarvestAction and adds it as a possible move for the player). The execute method in the HarvestAction class calls the harvestFruit method in Tree (which has a 40% chance of returning a Fruit object).