Create Corpse class to represent dead dinosaurs
Previously, corpses were objects of the PortableItem class. This meant that the player could illogically pick up a dead dinosaur. This would also make it difficult for allosaurs to detect if they are adjacent to a corpse (that they can eat) or not.
This change was incorporated into the existing AttackAction class and stegosaurs-dying-from-hunger functionality.