Create Carnivore, Herbivore and Omnivore classes (which extend Dinosaur)
Now, Stegosaur extends Herbivore and Allosaur extends Carnivore. This makes the system more extensible, in that there is no longer a need to duplicate, for example, the lookForFoodBehaviour method in Allosaur (for Archaeopteryx); Allosaur and Archaeopteryx are both carnivores and can therefore share the same method. This also allows any further behaviour, that is specific to a dinosaur's diet, to be added with ease.